# docs/openapi/openapi_combined.yaml
openapi: 3.0.3
info:
  title: Social Media API
  description: Объединенная документация всех микросервисов
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: API Gateway (Development)
paths:
...
  /api/posts:
    post:
      tags:
      - posts
      summary: Create Post
      operationId: create_post_api_posts_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - posts
      summary: List Posts
      operationId: list_posts_api_posts_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 10
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPostOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/posts/{post_id}:
    get:
      tags:
      - posts
      summary: Get Post
      operationId: get_post_api_posts__post_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - posts
      summary: Update Post
      operationId: update_post_api_posts__post_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - posts
      summary: Delete Post
      operationId: delete_post_api_posts__post_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/posts/{post_id}/view:
    post:
      tags:
      - posts
      summary: View Post
      operationId: view_post_api_posts__post_id__view_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ViewPostResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/posts/{post_id}/like:
    post:
      tags:
      - posts
      summary: Like Post
      operationId: like_post_api_posts__post_id__like_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LikePostResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - posts
      summary: Unlike Post
      operationId: unlike_post_api_posts__post_id__like_delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LikePostResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/posts/{post_id}/comments:
    post:
      tags:
      - posts
      summary: Comment Post
      operationId: comment_post_api_posts__post_id__comments_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - posts
      summary: Get Post Comments
      operationId: get_post_comments_api_posts__post_id__comments_get
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 10
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCommentsOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    CommentCreate:
      properties:
        content:
          type: string
          minLength: 1
          title: Content
          description: Comment content cannot be empty
      type: object
      required:
      - content
      title: CommentCreate
    CommentOut:
      properties:
        id:
          type: integer
          title: Id
        post_id:
          type: string
          title: Post Id
        user_id:
          type: integer
          title: User Id
        content:
          type: string
          title: Content
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - post_id
      - user_id
      - content
      - created_at
      - updated_at
      title: CommentOut
    CommentResponse:
      properties:
        success:
          type: boolean
          title: Success
        message:
          type: string
          title: Message
        comment:
          anyOf:
          - $ref: '#/components/schemas/CommentOut'
          - type: 'null'
      type: object
      required:
      - success
      - message
      title: CommentResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LikePostResponse:
      properties:
        success:
          type: boolean
          title: Success
        message:
          type: string
          title: Message
        total_likes:
          type: integer
          title: Total Likes
      type: object
      required:
      - success
      - message
      - total_likes
      title: LikePostResponse
    PaginatedCommentsOut:
      properties:
        comments:
          items:
            $ref: '#/components/schemas/CommentOut'
          type: array
          title: Comments
        total_count:
          type: integer
          title: Total Count
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
      - comments
      - total_count
      - page
      - page_size
      title: PaginatedCommentsOut
    PaginatedPostOut:
      properties:
        posts:
          items:
            $ref: '#/components/schemas/PostOut'
          type: array
          title: Posts
        total_count:
          type: integer
          title: Total Count
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
      - posts
      - total_count
      - page
      - page_size
      title: PaginatedPostOut
    PostCreate:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        is_private:
          type: boolean
          title: Is Private
          default: false
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
          default: []
      type: object
      required:
      - title
      - description
      title: PostCreate
...
    PostOut:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        author_id:
          type: integer
          title: Author Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        is_private:
          type: boolean
          title: Is Private
        tags:
          items:
            type: string
          type: array
          title: Tags
      type: object
      required:
      - id
      - title
      - description
      - author_id
      - created_at
      - updated_at
      - is_private
      - tags
      title: PostOut
...
    PostUpdate:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        is_private:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Private
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
      type: object
      title: PostUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ViewPostResponse:
      properties:
        success:
          type: boolean
          title: Success
        message:
          type: string
          title: Message
      type: object
      required:
      - success
      - message
      title: ViewPostResponse